# gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Todo
    description: Todo-Backend Implementation with GOLANG & GOA
    version: 0.0.1
servers:
    - url: http://localhost:8080
paths:
    /todos:
        delete:
            tags:
                - todos
            summary: deleteAll todos
            operationId: todos#deleteAll
            responses:
                "204":
                    description: No Content response.
        get:
            tags:
                - todos
            summary: list todos
            operationId: todos#list
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Todo'
                                example:
                                    - completed: false
                                      id: 6616090331682594200
                                      order: 8470419619838530899
                                      title: Vel libero debitis.
                                      url: Facere aperiam minima mollitia accusantium ut quis.
                                    - completed: false
                                      id: 6616090331682594200
                                      order: 8470419619838530899
                                      title: Vel libero debitis.
                                      url: Facere aperiam minima mollitia accusantium ut quis.
                                    - completed: false
                                      id: 6616090331682594200
                                      order: 8470419619838530899
                                      title: Vel libero debitis.
                                      url: Facere aperiam minima mollitia accusantium ut quis.
                            example:
                                - completed: false
                                  id: 6616090331682594200
                                  order: 8470419619838530899
                                  title: Vel libero debitis.
                                  url: Facere aperiam minima mollitia accusantium ut quis.
                                - completed: false
                                  id: 6616090331682594200
                                  order: 8470419619838530899
                                  title: Vel libero debitis.
                                  url: Facere aperiam minima mollitia accusantium ut quis.
                                - completed: false
                                  id: 6616090331682594200
                                  order: 8470419619838530899
                                  title: Vel libero debitis.
                                  url: Facere aperiam minima mollitia accusantium ut quis.
                                - completed: false
                                  id: 6616090331682594200
                                  order: 8470419619838530899
                                  title: Vel libero debitis.
                                  url: Facere aperiam minima mollitia accusantium ut quis.
        post:
            tags:
                - todos
            summary: create todos
            operationId: todos#create
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TodoPayload'
                        example:
                            completed: false
                            order: 3935193967633306166
                            title: Voluptatibus adipisci eos vel.
            responses:
                "201":
                    description: Created response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Todo'
                            example:
                                completed: true
                                id: 4753212596655035884
                                order: 7041713459340774230
                                title: Quia rerum enim cupiditate eaque nihil dolore.
                                url: Voluptatum molestias assumenda.
                "400":
                    description: 'BadRequest: Invalid user input.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
    /todos/{id}:
        delete:
            tags:
                - todos
            summary: delete todos
            operationId: todos#delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                    example: Eligendi dolor et maiores autem.
                  example: Esse nisi rerum a ipsam delectus qui.
            responses:
                "204":
                    description: No Content response.
                "404":
                    description: 'NotFound: Todo not found.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
        get:
            tags:
                - todos
            summary: read todos
            operationId: todos#read
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                    example: Corrupti autem occaecati.
                  example: Qui ipsa molestiae delectus quis.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Todo'
                            example:
                                completed: false
                                id: 6048023540673081850
                                order: 5165791562111512251
                                title: Dolor quia et est.
                                url: Et quasi.
                "404":
                    description: 'NotFound: Todo not found.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
        patch:
            tags:
                - todos
            summary: update todos
            operationId: todos#update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                    example: Iure corrupti.
                  example: Culpa nobis.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateRequestBody'
                        example:
                            completed: true
                            order: 4610252795505360349
                            title: Molestiae molestias rem exercitationem vel officia consequatur.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Todo'
                            example:
                                completed: false
                                id: 6834463199987684273
                                order: 5648491383675271446
                                title: Dolorum cupiditate est error magni fugit.
                                url: At aut vero.
                "400":
                    description: 'BadRequest: Invalid user input.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "404":
                    description: 'NotFound: Todo not found.'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
components:
    schemas:
        Error:
            type: object
            properties:
                fault:
                    type: boolean
                    description: Is the error a server-side fault?
                    example: true
                id:
                    type: string
                    description: ID is a unique identifier for this particular occurrence of the problem.
                    example: 123abc
                message:
                    type: string
                    description: Message is a human-readable explanation specific to this occurrence of the problem.
                    example: parameter 'p' must be an integer
                name:
                    type: string
                    description: Name is the name of this class of errors.
                    example: bad_request
                temporary:
                    type: boolean
                    description: Is the error temporary?
                    example: false
                timeout:
                    type: boolean
                    description: Is the error a timeout?
                    example: true
            description: Invalid user input.
            example:
                fault: true
                id: 123abc
                message: parameter 'p' must be an integer
                name: bad_request
                temporary: true
                timeout: false
            required:
                - name
                - id
                - message
                - temporary
                - timeout
                - fault
        Todo:
            type: object
            properties:
                completed:
                    type: boolean
                    description: Whether the task is completed or not
                    example: false
                id:
                    type: integer
                    description: Unique ID
                    example: 6347800286872210480
                    format: int64
                order:
                    type: integer
                    description: Order or priority of the task
                    example: 695485723255361810
                    format: int64
                title:
                    type: string
                    description: Title of the task
                    example: Repellat culpa ipsum optio at doloribus ea.
                url:
                    type: string
                    example: Culpa quisquam assumenda.
            example:
                completed: false
                id: 9133958982753990109
                order: 5548886307755879753
                title: Eum ipsum quam recusandae fuga velit.
                url: Recusandae corporis et sed.
            required:
                - id
                - url
        TodoPayload:
            type: object
            properties:
                completed:
                    type: boolean
                    description: Whether the task is completed or not
                    example: true
                order:
                    type: integer
                    description: Order or priority of the task
                    example: 7958389109286441471
                    format: int64
                title:
                    type: string
                    description: Title of the task
                    example: Voluptatem qui assumenda ipsum.
            example:
                completed: true
                order: 6049171131321094305
                title: Non assumenda aliquid perspiciatis voluptatem ut.
        UpdateRequestBody:
            type: object
            properties:
                completed:
                    type: boolean
                    description: Whether the task is completed or not
                    example: true
                order:
                    type: integer
                    description: Order or priority of the task
                    example: 7706733665398639614
                    format: int64
                title:
                    type: string
                    description: Title of the task
                    example: Dolores quis rerum laudantium voluptatem magni.
            example:
                completed: false
                order: 4953306782095514376
                title: Distinctio facilis sint blanditiis aut.
tags:
    - name: todos